fix(customer-profile-board): fix UT error. (#204)
