+++ src/components/customer-profile-board/__tests__/test-queryStringSchema.js
@@ -13,7 +13,8 @@ describe('queryStringSchema', () => {
 		const customerInformation = {
 			nickName: 's_peggy',
 			shopId: '62847409',
-			platName: 'taobao'
+			platName: 'taobao',
+			tenantId: 'dagouzi01'
 		};
 
 		const queryString = `{
@@ -48,6 +49,53 @@ describe('queryStringSchema', () => {
           buyer_credit_lev
           mobile
         }
+	    }
+	  }
+
+	  rfm:shuyun_searchapi_rfm(
+	    param: {
+	  settings:{
+	    data_source: "rfm"
+	    query_id:"report-id"
+	    pagination:{
+	      limit: 10
+	      offset:0
+	    }
+	    return_format:"json"
+	  }
+	  filters:{
+	    type:"and"
+	    fields: [
+	      { type:"str_selector" dimension:"dp_id" value:"62847409" }
+	      { type:"str_selector" dimension:"buyer_nick" value:"s_peggy" }
+	    ]
+	  }
+	}
+	  ){
+	    flag
+	    msg
+	    data{
+	      data{
+	        period
+	        trade_first_time
+	        trade_first_amount
+	        trade_first_interval
+	        trade_last_time
+	        trade_last_amount
+	        trade_last_interval
+	        trade_tidcount
+	        trade_count
+	        trade_amount
+	        trade_item_num
+	        trade_avg_amount
+	        trade_avg_item_num
+	        trade_avg_buy_interval
+	        trade_refund_count
+	        trade_refund_amount
+	        trade_avg_confirm_interval
+	        trade_max_amount
+	        trade_order_discount_fee
+	      }
 	    }
 	  }
 
@@ -126,8 +174,22 @@ describe('queryStringSchema', () => {
 			sinawbUserName
 		}
 
+		custom_property_customer(
+      customerNo:"s_peggy"
+      tenantId:"dagouzi01"
+      platform:"taobao"
+    ){
+      properties{
+        id
+        name
+        type
+        optional
+        value
+      }
+    }
+
 		}`;
 
-		assert.deepEqual(generatorQueryString(customerInformation.nickName, customerInformation.shopId, customerInformation.platName).replace(/\s/g, ''), queryString.replace(/\s/g, ''));
+		assert.deepEqual(generatorQueryString(customerInformation.nickName, customerInformation.shopId, customerInformation.platName, customerInformation.tenantId).replace(/\s/g, ''), queryString.replace(/\s/g, ''));
 	});
 });
